feat(email): implement resend verification email functionality and update token expiration
- Added endpoint/controller for resending verification email.
- Updated logic to generate a new verification token when user requests resend.
- Set token expiration to 30 minutes (previously 60 minutes).
- Integrated Nodemailer service to reuse the same email template.
- Added frontend button “Resend Email” with disabled timer to prevent spam.
- Enhanced backend validation to check if user is already verified before resending.

+++ backend/src/controller/userController.js
@@ -77,13 +77,43 @@ export const verifyToken = async (req, res) => {
         await User.updateOne({ _id: id }, { isVerified: true });
 
         //deleting the token in redis
-        deleteToken(id)
+        await deleteToken(id)
 
         res.status(200).redirect(`http://localhost:5173/verify/email/link?status=success&email=${encodeURIComponent(user.email)}`)
 
     } catch (error) {
         res.status(500).json({ success: false, message: "internal server error" })
     }
+}
 
+export const resendVerificationMail = async (req, res) => {
 
+    try {
+
+        const { email } = req.body
+
+        if (!email)
+            return res.status(400).json({ success: false, message: "Email not found! " })
+
+        const isExisting = await User.findOne({ email })
+
+        if (!isExisting)
+            return res.status(400).json({ success: false, message: "Email is Invalid!" })
+
+        const id = isExisting._id.toString()
+
+        await deleteToken(id)
+
+        const token = await generateAndStoreToken(id)
+
+        //basic skeleton of the verification link send to user
+        const verificationLink = `http://localhost:5000/api/user/verify-email/${id}/${token}`;
+
+        await sendVerificationEmail(email, verificationLink)
+        return res.status(200).json({ success: true, message: "verification email send!" })
+
+    } catch (error) {
+        console.log("Error while resending email", error)
+        res.status(500).json({ success: false, message: "Error while resending email" })
+    }
 }
